pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auditrum"
version = "0.3.0"
description = "Auditrum — a PostgreSQL audit system with rich contextual logging and Django integration"
readme = "README.md"
authors = [
    {name = "Andrei Karbanovich", email = "andrey.karbanovich@icloud.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

keywords = ["postgresql", "audit", "django", "database", "triggers", "history", "track changes"]

# Core dependencies (for CLI and audit logic)
dependencies = [
    "typer>=0.9",
    "psycopg[binary]>=3.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "structlog>=23.0",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
django = [
    "Django>=4.2",
]
sqlalchemy = [
    "SQLAlchemy>=2.0",
]
observability = [
    "opentelemetry-api>=1.20",
    "prometheus-client>=0.17",
    "sentry-sdk>=1.30",
]
test = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "testcontainers[postgres]>=4.0",
]

[project.urls]
Homepage = "https://github.com/tauvin/auditrum"
BugTracker = "https://github.com/tauvin/auditrum/issues"
Documentation = "https://auditrum.readthedocs.io/"

[project.scripts]
auditrum = "auditrum.cli:app"

[tool.setuptools.packages.find]
include = ["auditrum*"]

[tool.setuptools.package-data]
auditrum = ["**/*.html", "**/*.txt", "**/*.sql"]
